🏗️ Flatten the `src/redux/` folder structure
Since error page management doesn't belong to any specific feature, and the folder-by-type approach is no longer considered a good Redux practice, this folder structure needed to be flattened.
