--- src/redux/store/index.js
@@ -1,19 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit'
-import { userReducer } from '../../features/user/store/userSlice'
-import { errorPage } from '../reducers/errorPage'
-import { postsReducer } from '../../features/activity-feed/post/store/postsSlice'
-import { commentsReducer } from '../../features/activity-feed/comment/store/commentsSlice'
-import { reactionsReducer } from '../../features/activity-feed/reaction/store/reactionsSlice'
-
-const reducer = {
-    user: userReducer,
-    errorPage,
-    posts: postsReducer,
-    comments: commentsReducer,
-    reactions: reactionsReducer
-}
-
-export const store = configureStore({
-    reducer,
-    devTools: true
-})
